This exposes the firrtl testing utils to repos that have a dependency on firrtl. These functions are in identically reproduced in chisel3 src/test/scala
